♻️ observer 역할을 hook으로 분리
