+++ client/src/views/room-view.tsx
@@ -1,16 +1,17 @@
 /* eslint-disable max-len */
 /* eslint-disable no-underscore-dangle */
 import React, {
-  MouseEvent, useEffect, useRef, useState,
+  MouseEvent, useEffect, useState,
 } from 'react';
 import styled from 'styled-components';
-import { useRecoilState, useSetRecoilState } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 
 import RoomCard from '@common/room-card';
 import useFetchItems from '@src/hooks/useFetchItems';
 import LoadingSpinner from '@common/loading-spinner';
 import roomViewType from '@atoms/room-view-type';
 import roomDocumentIdState from '@atoms/room-document-id';
+import useItemFecthObserver from '@src/hooks/useItemFetchObserver';
 import { nowFetchingState } from '@src/recoil/atoms/main-section-scroll';
 
 interface Participants{
@@ -57,9 +58,9 @@ export function RoomCardList({ roomList, roomCardClickHandler }:
 
 function RoomView() {
   const [nowItemList, nowItemType] = useFetchItems<RoomCardProps>('/room', 'room');
-  const [nowFetching, setNowFetching] = useRecoilState(nowFetchingState);
-  const targetRef = useRef<HTMLDivElement>(null);
+  const nowFetching = useRecoilValue(nowFetchingState);
   const [loading, setLoading] = useState(true);
+  const [targetRef] = useItemFecthObserver(loading);
   const setRoomView = useSetRecoilState(roomViewType);
   const setRoomDocumentId = useSetRecoilState(roomDocumentIdState);
 
@@ -70,23 +71,6 @@ function RoomView() {
     if (roomDocumentId) setRoomDocumentId(roomDocumentId);
     else console.error('no room-id');
   };
-
-  const onIntersect = async (entries: IntersectionObserverEntry[]) => {
-    if (entries[0].isIntersecting && !nowFetching) {
-      setNowFetching(true);
-    }
-  };
-
-  useEffect(() => {
-    let observer: IntersectionObserver;
-    if (targetRef.current) {
-      observer = new IntersectionObserver(onIntersect, {
-        threshold: 0.4,
-      });
-      observer.observe(targetRef.current);
-    }
-    return () => observer?.disconnect();
-  }, [targetRef.current, loading]);
 
   useEffect(() => {
     if (nowItemList && nowItemType === 'room') {
